move skipped test to main tests
